requirements/local.txt
Pillow
asgiref==3.8.1
Django==5.1
django-model-utils==4.5.1
djangorestframework==3.16.1
pytz==2024.1
sqlparse==0.5.1
Unipath==1.1